py3 support! yeah

+++ fancytools/stats/inverseHistogram.py
@@ -1,3 +1,5 @@
+from __future__ import division
+from builtins import zip
 import numpy as np
 
 
@@ -9,12 +11,12 @@ def inverseHistogram(hist, bin_range):
         np.array: data that would create the same histogram as given
     '''
     data = hist.astype(float) / np.min(hist[np.nonzero(hist)])
-    new_data = np.empty(shape=np.sum(data))
+    new_data = np.empty(shape=np.sum(data, dtype=int))
     i = 0
     xvals = np.linspace(bin_range[0], bin_range[1],len(data))
     for d, x in zip(data, xvals):
         new_data[i:i+d] = x
-        i += d
+        i += int(d)
     return new_data
 
 
@@ -40,4 +42,4 @@ if __name__ == '__main__':
         plt.figure(2)
         plt.plot(data)
         plt.plot(data2)
-    plt.show()
+        plt.show()
